requirements.txt
Django==1.5.1
MySQL-python==1.2.3
South==0.8.2
amqp==1.0.13
anyjson==0.3.3
argparse==1.2.1
billiard==2.7.3.31
django-tastypie==0.10.0
mimeparse==0.1.3
